requirements.txt
tqdm
pillow
pyyaml
requests
cvat_sdk
